+++ dbdaora/boolean/_tests/datastore/test_integration_service_boolean_aioredis_datastore_get_many.py
@@ -49,7 +49,6 @@ async def test_should_get_many_from_cache(
 async def test_should_get_many_from_fallback_when_not_found_on_memory(
     fake_service, serialized_fake_entity, fake_entity, fake_entity2
 ):
-    client = fake_service.repository.fallback_data_source.client
     await fake_service.repository.memory_data_source.delete(
         'fake:other_fake:fake'
     )
@@ -63,10 +62,16 @@ async def test_should_get_many_from_fallback_when_not_found_on_memory(
         'fake:not-found:other_fake:fake2'
     )
     await fake_service.repository.fallback_data_source.put(
-        client.key('fake', 'other_fake:fake'), {'value': True}
+        fake_service.repository.fallback_data_source.make_key(
+            'fake', 'other_fake:fake'
+        ),
+        {'value': True},
     )
     await fake_service.repository.fallback_data_source.put(
-        client.key('fake', 'other_fake:fake2'), {'value': True},
+        fake_service.repository.fallback_data_source.make_key(
+            'fake', 'other_fake:fake2'
+        ),
+        {'value': True},
     )
 
     entities = [