+++ dbdaora/hash/_tests/datastore/test_integration_service_hash_aioredis_datastore_get_many.py
@@ -62,7 +62,6 @@ async def test_should_get_many_from_cache(
 async def test_should_get_many_from_fallback_when_not_found_on_memory(
     fake_service, serialized_fake_entity, fake_entity, fake_entity2
 ):
-    client = fake_service.repository.fallback_data_source.client
     await fake_service.repository.memory_data_source.delete(
         'fake:other_fake:fake'
     )
@@ -76,10 +75,15 @@ async def test_should_get_many_from_fallback_when_not_found_on_memory(
         'fake:not-found:other_fake:fake2'
     )
     await fake_service.repository.fallback_data_source.put(
-        client.key('fake', 'other_fake:fake'), dataclasses.asdict(fake_entity)
+        fake_service.repository.fallback_data_source.make_key(
+            'fake', 'other_fake:fake'
+        ),
+        dataclasses.asdict(fake_entity),
     )
     await fake_service.repository.fallback_data_source.put(
-        client.key('fake', 'other_fake:fake2'),
+        fake_service.repository.fallback_data_source.make_key(
+            'fake', 'other_fake:fake2'
+        ),
         dataclasses.asdict(fake_entity2),
     )
 
@@ -103,7 +107,6 @@ async def test_should_get_many_from_fallback_when_not_found_on_memory(
 async def test_should_get_many_from_fallback_when_not_found_on_memory_with_fields(
     fake_service, serialized_fake_entity, fake_entity, fake_entity2
 ):
-    client = fake_service.repository.fallback_data_source.client
     await fake_service.repository.memory_data_source.delete(
         'fake:other_fake:fake'
     )
@@ -117,10 +120,15 @@ async def test_should_get_many_from_fallback_when_not_found_on_memory_with_field
         'fake:not-found:other_fake:fake2'
     )
     await fake_service.repository.fallback_data_source.put(
-        client.key('fake', 'other_fake:fake'), dataclasses.asdict(fake_entity)
+        fake_service.repository.fallback_data_source.make_key(
+            'fake', 'other_fake:fake'
+        ),
+        dataclasses.asdict(fake_entity),
     )
     await fake_service.repository.fallback_data_source.put(
-        client.key('fake', 'other_fake:fake2'),
+        fake_service.repository.fallback_data_source.make_key(
+            'fake', 'other_fake:fake2'
+        ),
         dataclasses.asdict(fake_entity2),
     )
     fake_entity.number = None