+++ dbdaora/hash/_tests/datastore/test_integration_service_hash_aioredis_datastore_get_one.py
@@ -59,7 +59,6 @@ async def test_should_get_one_from_cache(
 async def test_should_get_one_from_fallback_when_not_found_on_memory(
     fake_service, serialized_fake_entity, fake_entity
 ):
-    client = fake_service.repository.fallback_data_source.client
     await fake_service.repository.memory_data_source.delete(
         'fake:other_fake:fake'
     )
@@ -67,7 +66,10 @@ async def test_should_get_one_from_fallback_when_not_found_on_memory(
         'fake:not-found:other_fake:fake'
     )
     await fake_service.repository.fallback_data_source.put(
-        client.key('fake', 'other_fake:fake'), dataclasses.asdict(fake_entity)
+        fake_service.repository.fallback_data_source.make_key(
+            'fake', 'other_fake:fake'
+        ),
+        dataclasses.asdict(fake_entity),
     )
 
     entity = await fake_service.get_one('fake', other_id='other_fake')
@@ -82,12 +84,14 @@ async def test_should_get_one_from_fallback_when_not_found_on_memory(
 async def test_should_get_one_from_fallback_when_not_found_on_memory_with_fields(
     fake_service, serialized_fake_entity, fake_entity
 ):
-    client = fake_service.repository.fallback_data_source.client
     await fake_service.repository.memory_data_source.delete(
         'fake:other_fake:fake'
     )
     await fake_service.repository.fallback_data_source.put(
-        client.key('fake', 'other_fake:fake'), dataclasses.asdict(fake_entity)
+        fake_service.repository.fallback_data_source.make_key(
+            'fake', 'other_fake:fake'
+        ),
+        dataclasses.asdict(fake_entity),
     )
     fake_entity.number = None
     fake_entity.boolean = None