+++ dbdaora/hash/_tests/mongodb/test_integration_service_hash_aioredis_mongodb_get_many_composed_key.py
@@ -6,7 +6,6 @@ from aioredis import RedisError
 from jsondaora import dataclasses
 
 from dbdaora import HashRepository
-from dbdaora.data_sources.fallback.mongodb import Key
 
 
 @pytest.fixture
@@ -88,10 +87,16 @@ async def test_should_get_many_from_fallback_when_not_found_on_memory(
         'fake:not-found:other_fake:fake2'
     )
     await fake_service.repository.fallback_data_source.put(
-        Key('fake', 'other_fake:fake'), dataclasses.asdict(fake_entity)
+        fake_service.repository.fallback_data_source.make_key(
+            'fake', 'other_fake:fake'
+        ),
+        dataclasses.asdict(fake_entity),
     )
     await fake_service.repository.fallback_data_source.put(
-        Key('fake', 'other_fake:fake2'), dataclasses.asdict(fake_entity2),
+        fake_service.repository.fallback_data_source.make_key(
+            'fake', 'other_fake:fake2'
+        ),
+        dataclasses.asdict(fake_entity2),
     )
 
     entities = [
@@ -127,10 +132,16 @@ async def test_should_get_many_from_fallback_when_not_found_on_memory_with_field
         'fake:not-found:other_fake:fake2'
     )
     await fake_service.repository.fallback_data_source.put(
-        Key('fake', 'other_fake:fake'), dataclasses.asdict(fake_entity)
+        fake_service.repository.fallback_data_source.make_key(
+            'fake', 'other_fake:fake'
+        ),
+        dataclasses.asdict(fake_entity),
     )
     await fake_service.repository.fallback_data_source.put(
-        Key('fake', 'other_fake:fake2'), dataclasses.asdict(fake_entity2),
+        fake_service.repository.fallback_data_source.make_key(
+            'fake', 'other_fake:fake2'
+        ),
+        dataclasses.asdict(fake_entity2),
     )
     fake_entity.number = None
     fake_entity.boolean = None